moved benchmarking tests into separate module
