--- app-norx/src/androidTest/java/com/alterego/stackoverflow/norx/test/di/AndroidTestProvidersModule.java
@@ -1,14 +1,0 @@
-package com.alterego.stackoverflow.norx.test.di;
-
-import android.app.Application;
-
-import dagger.Module;
-
-@Module
-public class AndroidTestProvidersModule extends AndroidModule {
-
-    public AndroidTestProvidersModule(Application app) {
-        super(app);
-    }
-
-}
