--- app-norx/src/androidTest/java/com/alterego/stackoverflow/norx/test/rules/Repeat.java
@@ -1,14 +1,0 @@
-package com.alterego.stackoverflow.norx.test.rules;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({
-        java.lang.annotation.ElementType.METHOD
-})
-public @interface Repeat {
-
-    int times();
-}
