+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':app-norx'
+include ':app', ':app-norx', ':testapplication'
