+++ testapplication/src/androidTest/java/com/alterego/stackoverflow/test/api/StackOverflowApiResponses.java
@@ -1,0 +1,791 @@
+package com.alterego.stackoverflow.test.api;
+
+public class StackOverflowApiResponses {
+
+    public static final String SEARCH_RAW_RESPONSE = "{\n"
+            + "    \"statusCode\" : 200,\n"
+            + "    \"headers\": {\n"
+            + "        \"Cache-Control\": \"no-transform,public,max-age=1\",\n"
+            + "        \"Content-Type\": \" application/json\",\n"
+            + "        \"Etag\": \"1416996436511\"\n"
+            + "    },\n"
+            + "    \"body\": {\n"
+            + "  \"items\": [\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"reactive-programming\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 368,\n"
+            + "        \"user_id\": 1720101,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 38,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/85ae256b19b6e9d19d06dd704d24f917?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"Shvalb\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1720101/shvalb\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 27,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1468102424,\n"
+            + "      \"creation_date\": 1468036469,\n"
+            + "      \"question_id\": 38278132,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38278132/delay-between-iterated-elements-and-on-subscribe\",\n"
+            + "      \"title\": \"Delay between iterated elements and on subscribe\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 52,\n"
+            + "        \"user_id\": 6348948,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 100,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/d14111d5f51d0b014c37fd5dd63bacec?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"otal\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/6348948/otal\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 24,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 2,\n"
+            + "      \"last_activity_date\": 1468099762,\n"
+            + "      \"creation_date\": 1468084368,\n"
+            + "      \"question_id\": 38284285,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38284285/schedulers-io-on-parallel-database-search\",\n"
+            + "      \"title\": \"Schedulers.io() on parallel database search\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"architecture\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"mvp\",\n"
+            + "        \"retrofit2\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 811,\n"
+            + "        \"user_id\": 1651286,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/WvZb0.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"Akhil Dad\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1651286/akhil-dad\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 50,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 2,\n"
+            + "      \"last_activity_date\": 1468094736,\n"
+            + "      \"creation_date\": 1467885177,\n"
+            + "      \"question_id\": 38242462,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38242462/internet-check-where-to-place-when-using-mvp-rx-and-retrofit\",\n"
+            + "      \"title\": \"Internet check, where to place when using MVP, RX and Retrofit\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"dependency-injection\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"google-api-client\",\n"
+            + "        \"dagger-2\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 92,\n"
+            + "        \"user_id\": 4515136,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 67,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/593d41322fdf9856768987dc5c213033?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"buddhabath\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/4515136/buddhabath\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 41,\n"
+            + "      \"bounty_amount\": 50,\n"
+            + "      \"bounty_closes_date\": 1468679348,\n"
+            + "      \"answer_count\": 0,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1468074696,\n"
+            + "      \"creation_date\": 1467891940,\n"
+            + "      \"last_edit_date\": 1468074696,\n"
+            + "      \"question_id\": 38244651,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38244651/googleapiclient-must-not-be-null-awareness-api\",\n"
+            + "      \"title\": \"GoogleApiClient must not be null [Awareness API]\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 91,\n"
+            + "        \"user_id\": 4519179,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://lh6.googleusercontent.com/-B9R2KsHIbEU/AAAAAAAAAAI/AAAAAAAAAEA/RHNhzZPaRVs/photo.jpg?sz=128\",\n"
+            + "        \"display_name\": \"Debu\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/4519179/debu\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 44,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1468047529,\n"
+            + "      \"creation_date\": 1467720788,\n"
+            + "      \"last_edit_date\": 1467733716,\n"
+            + "      \"question_id\": 38203152,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38203152/detecting-value-change-using-rxjava\",\n"
+            + "      \"title\": \"Detecting value change using Rxjava\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"retrofit\",\n"
+            + "        \"realm\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"retrofit2\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 35,\n"
+            + "        \"user_id\": 6338257,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/8996eeff3e82d52740ed5b9a3bd98565?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"shilch\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/6338257/shilch\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 20,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1468015794,\n"
+            + "      \"creation_date\": 1468011218,\n"
+            + "      \"question_id\": 38275247,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38275247/disabling-networking-stops-emitting-data-from-realm\",\n"
+            + "      \"title\": \"Disabling networking stops emitting data from realm\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"bluetooth-lowenergy\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-android\",\n"
+            + "        \"rxandroidble\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 1643,\n"
+            + "        \"user_id\": 183362,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 80,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/22fe5ae8af7189ac8276be26f5d9e1f2?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"Tim Clemons\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/183362/tim-clemons\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 104,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467992329,\n"
+            + "      \"creation_date\": 1466616524,\n"
+            + "      \"last_edit_date\": 1466631773,\n"
+            + "      \"question_id\": 37974679,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/37974679/managing-side-effects-with-multiple-rxjava-observables\",\n"
+            + "      \"title\": \"Managing side-effects with multiple RxJava Observables\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-android\",\n"
+            + "        \"rxandroidble\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 1,\n"
+            + "        \"user_id\": 6560826,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/54116132948e9fe50262293b6547d07c?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"KolinLoures\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/6560826/kolinloures\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 29,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467991122,\n"
+            + "      \"creation_date\": 1467908417,\n"
+            + "      \"question_id\": 38250610,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38250610/sending-the-list-of-commands-to-device-using-rxandroidble-rxjava\",\n"
+            + "      \"title\": \"Sending the list of commands to device using RxAndroidBle (rxJava)\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"java-7\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 2235,\n"
+            + "        \"user_id\": 587467,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 81,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/7787bdb4718fdc9b42fb46d671bd7eca?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"Tar\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/587467/tar\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 34,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467984773,\n"
+            + "      \"creation_date\": 1467979318,\n"
+            + "      \"question_id\": 38266147,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38266147/observable-createsynconsubscribe-createstateless-throws-illegalstateexcep\",\n"
+            + "      \"title\": \"Observable.create(SyncOnSubscribe.createStateless(...)) throws IllegalStateException: &quot;onNext called multiple times!&quot;\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-android\",\n"
+            + "        \"reactivex\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 239,\n"
+            + "        \"user_id\": 1759005,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 50,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/d7024554a43616d15116e04a559c8c70?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"jpmastermind\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1759005/jpmastermind\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 40,\n"
+            + "      \"accepted_answer_id\": 38261903,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467968190,\n"
+            + "      \"creation_date\": 1467959072,\n"
+            + "      \"question_id\": 38260171,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38260171/rxandroid-observe-an-edittext-while-also-observing-touches-on-2-views\",\n"
+            + "      \"title\": \"RxAndroid: Observe an EditText while also observing touches on 2 Views\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 576,\n"
+            + "        \"user_id\": 1150503,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 65,\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/ZzHBn.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"Dariusz Mydlarz\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1150503/dariusz-mydlarz\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 25,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467962619,\n"
+            + "      \"creation_date\": 1467961385,\n"
+            + "      \"question_id\": 38260733,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38260733/rxjava-how-to-complete-observable-from-a-different-place\",\n"
+            + "      \"title\": \"RxJava: how to complete observable from a different place?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"json\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"retrofit2\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 61,\n"
+            + "        \"user_id\": 2990544,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/867dcdf7bf59fb37006e006026535d8b?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"Thilina Anuradh\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/2990544/thilina-anuradh\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 13,\n"
+            + "      \"answer_count\": 0,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467955305,\n"
+            + "      \"creation_date\": 1467955305,\n"
+            + "      \"question_id\": 38259373,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38259373/retrofit-accumilative-json-uploads\",\n"
+            + "      \"title\": \"Retrofit Accumilative JSON uploads\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\",\n"
+            + "        \"take-until\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 6102,\n"
+            + "        \"user_id\": 296427,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 87,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/8692a55fc3524ed4e7cb85b0b3b4693c?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"pixel\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/296427/pixel\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 47,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467920249,\n"
+            + "      \"creation_date\": 1467631281,\n"
+            + "      \"question_id\": 38183494,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38183494/rxjava-takeuntil-with-emmision-of-last-item\",\n"
+            + "      \"title\": \"RxJava takeUntil with emmision of last item?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"system.reactive\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 2416,\n"
+            + "        \"user_id\": 471744,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 67,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/255e73e8f6e4e38dbbc060335c593b0c?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"spierce7\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/471744/spierce7\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 3040,\n"
+            + "      \"accepted_answer_id\": 21208440,\n"
+            + "      \"answer_count\": 3,\n"
+            + "      \"score\": 10,\n"
+            + "      \"last_activity_date\": 1467902952,\n"
+            + "      \"creation_date\": 1390027965,\n"
+            + "      \"question_id\": 21201083,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/21201083/rx-run-zipped-observables-in-parallel\",\n"
+            + "      \"title\": \"RX: Run Zipped Observables in parallel?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"gson\",\n"
+            + "        \"retrofit\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 22,\n"
+            + "        \"user_id\": 5262623,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/ea6543f7294c58d59b44689885321131?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"Even2015\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/5262623/even2015\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 29,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467894484,\n"
+            + "      \"creation_date\": 1467855363,\n"
+            + "      \"question_id\": 38236377,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38236377/how-to-explain-the-irregular-json-with-gson-rxjava-and-retrofit\",\n"
+            + "      \"title\": \"How to explain the irregular Json with Gson,Rxjava and Retrofit\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 15,\n"
+            + "        \"user_id\": 6504134,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 50,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/85e2707f7d00c925c6a6dc6abbd82e8d?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"Kordova\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/6504134/kordova\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 38,\n"
+            + "      \"accepted_answer_id\": 38244548,\n"
+            + "      \"answer_count\": 3,\n"
+            + "      \"score\": 2,\n"
+            + "      \"last_activity_date\": 1467891649,\n"
+            + "      \"creation_date\": 1467827393,\n"
+            + "      \"question_id\": 38230530,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38230530/stop-observable-interval-and-send-oncomplete\",\n"
+            + "      \"title\": \"stop Observable.interval and send onComplete\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"reactive-programming\",\n"
+            + "        \"rx-android\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 1276,\n"
+            + "        \"user_id\": 2069407,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 67,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/9cf04e8273f7868fedf1fe32e1a5ec4f?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"s-hunter\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/2069407/s-hunter\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 47,\n"
+            + "      \"accepted_answer_id\": 38234271,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467888891,\n"
+            + "      \"creation_date\": 1467840838,\n"
+            + "      \"question_id\": 38234152,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38234152/how-can-i-make-this-rxjava-zip-to-run-in-parallel\",\n"
+            + "      \"title\": \"How can I make this rxjava zip to run in parallel?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"google-app-engine\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 349,\n"
+            + "        \"user_id\": 1166952,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 60,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/c6a5ba48816157fd11c59b5a02a13a0b?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"ffgiraldez\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1166952/ffgiraldez\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 25,\n"
+            + "      \"accepted_answer_id\": 38242976,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467886647,\n"
+            + "      \"creation_date\": 1467796839,\n"
+            + "      \"question_id\": 38220450,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38220450/problems-threads-with-rxjava-and-appengine\",\n"
+            + "      \"title\": \"Problems threads with RxJava and AppEngine\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"multithreading\",\n"
+            + "        \"transactions\",\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 58,\n"
+            + "        \"user_id\": 1969746,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/FlQZi.png?s=128&g=1\",\n"
+            + "        \"display_name\": \"FloGz\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1969746/flogz\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 31,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 2,\n"
+            + "      \"last_activity_date\": 1467839952,\n"
+            + "      \"creation_date\": 1467820631,\n"
+            + "      \"question_id\": 38228428,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38228428/rxjava-changing-thread-after-concat-map\",\n"
+            + "      \"title\": \"RxJava changing thread after concat map\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-android\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 2877,\n"
+            + "        \"user_id\": 759007,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 84,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/cd27f10d8d60171d5a67306b7ccf94e0?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"Ziem\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/759007/ziem\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 3240,\n"
+            + "      \"answer_count\": 6,\n"
+            + "      \"score\": 5,\n"
+            + "      \"last_activity_date\": 1467831591,\n"
+            + "      \"creation_date\": 1426007579,\n"
+            + "      \"last_edit_date\": 1426008357,\n"
+            + "      \"question_id\": 28969995,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/28969995/how-to-ignore-error-and-continue-infinite-stream\",\n"
+            + "      \"title\": \"How to ignore error and continue infinite stream?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\",\n"
+            + "        \"hystrix\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 16,\n"
+            + "        \"user_id\": 2651885,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/xRDoL.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"Ranjit Shinde\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/2651885/ranjit-shinde\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 24,\n"
+            + "      \"answer_count\": 0,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467829278,\n"
+            + "      \"creation_date\": 1467829278,\n"
+            + "      \"question_id\": 38231084,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38231084/could-not-acquire-a-semaphore-for-execution-and-fallback-failed-exception\",\n"
+            + "      \"title\": \"Could not acquire a semaphore for execution and fallback failed Exception\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"network-programming\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"reactivex\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 8508,\n"
+            + "        \"user_id\": 446461,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 82,\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/9HaOK.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"mreichelt\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/446461/mreichelt\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 31,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 3,\n"
+            + "      \"last_activity_date\": 1467819045,\n"
+            + "      \"creation_date\": 1467817129,\n"
+            + "      \"question_id\": 38227253,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38227253/rxjava-call-network-multiple-times-with-increasing-delay-ignoring-errors\",\n"
+            + "      \"title\": \"RxJava: Call network multiple times with increasing delay, ignoring errors\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"retrofit\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-android\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 1051,\n"
+            + "        \"user_id\": 3212712,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 75,\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/eZV9f.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"mohax\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/3212712/mohax\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 83,\n"
+            + "      \"accepted_answer_id\": 38208524,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467814221,\n"
+            + "      \"creation_date\": 1467574450,\n"
+            + "      \"last_edit_date\": 1467814221,\n"
+            + "      \"question_id\": 38173653,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38173653/how-to-recreate-or-reset-cached-observable-used-with-retrofit-to-get-new-data\",\n"
+            + "      \"title\": \"How to recreate or reset cached Observable, used with Retrofit to get new data?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"retrofit2\",\n"
+            + "        \"rx-android\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 6743,\n"
+            + "        \"user_id\": 1919641,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 61,\n"
+            + "        \"profile_image\": \"https://i.stack.imgur.com/T6fy5.jpg?s=128&g=1\",\n"
+            + "        \"display_name\": \"mjosh\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1919641/mjosh\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 53,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467800243,\n"
+            + "      \"creation_date\": 1464860660,\n"
+            + "      \"question_id\": 37588208,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/37588208/retrofit-fetch-authorization-token-based-on-specific-error-message-using-rxandro\",\n"
+            + "      \"title\": \"Retrofit fetch authorization token based on specific error message using RxAndroid\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"swift\",\n"
+            + "        \"multithreading\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"rx-swift\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 1664,\n"
+            + "        \"user_id\": 1007522,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 78,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/c025d141e407a7db9a5af64e41bfd56f?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"user1007522\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/1007522/user1007522\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 57,\n"
+            + "      \"accepted_answer_id\": 37986481,\n"
+            + "      \"answer_count\": 4,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467798386,\n"
+            + "      \"creation_date\": 1466612333,\n"
+            + "      \"last_edit_date\": 1467798386,\n"
+            + "      \"question_id\": 37973445,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/37973445/does-the-order-of-subscribeon-and-observeon-matter\",\n"
+            + "      \"title\": \"Does the order of subscribeOn and observeOn matter?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"java\",\n"
+            + "        \"android\",\n"
+            + "        \"rx-java\",\n"
+            + "        \"okhttp\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 63,\n"
+            + "        \"user_id\": 5357773,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/81d372ad4bc32055911d3a8c7eb0a524?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"Santos Black\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/5357773/santos-black\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 1793,\n"
+            + "      \"accepted_answer_id\": 32687975,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 12,\n"
+            + "      \"last_activity_date\": 1467795095,\n"
+            + "      \"creation_date\": 1442813236,\n"
+            + "      \"last_edit_date\": 1465218958,\n"
+            + "      \"question_id\": 32687921,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/32687921/using-rxjava-and-okhttp\",\n"
+            + "      \"title\": \"Using RxJava and Okhttp\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\",\n"
+            + "        \"retrofit2\",\n"
+            + "        \"rx-android\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 63,\n"
+            + "        \"user_id\": 3034559,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 53,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/1efd76b87776b5bef1293cc86e0d32bc?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"user3034559\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/3034559/user3034559\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 33,\n"
+            + "      \"accepted_answer_id\": 38219537,\n"
+            + "      \"answer_count\": 1,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467794029,\n"
+            + "      \"creation_date\": 1467743289,\n"
+            + "      \"question_id\": 38210229,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38210229/how-to-continue-to-get-response-data-from-rest-api-even-if-http-status-code-is-n\",\n"
+            + "      \"title\": \"How to continue to get response data from REST API even if http status code is not 200?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\",\n"
+            + "        \"reactive-programming\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 5228,\n"
+            + "        \"user_id\": 990616,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 70,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/823118e69e4aaecbf04ff2961954a456?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"NathanZ\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/990616/nathanz\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 29,\n"
+            + "      \"accepted_answer_id\": 38212879,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467753441,\n"
+            + "      \"creation_date\": 1467751679,\n"
+            + "      \"question_id\": 38212442,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38212442/rxjava-how-to-take-the-first-element-of-a-list-and-return-it-as-observable\",\n"
+            + "      \"title\": \"RxJava - How to take the first element of a list and return it as Observable\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\",\n"
+            + "        \"reactive-programming\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 5228,\n"
+            + "        \"user_id\": 990616,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"accept_rate\": 70,\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/823118e69e4aaecbf04ff2961954a456?s=128&d=identicon&r=PG\",\n"
+            + "        \"display_name\": \"NathanZ\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/990616/nathanz\"\n"
+            + "      },\n"
+            + "      \"is_answered\": false,\n"
+            + "      \"view_count\": 36,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 0,\n"
+            + "      \"last_activity_date\": 1467739814,\n"
+            + "      \"creation_date\": 1467727977,\n"
+            + "      \"question_id\": 38205707,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/38205707/how-to-include-conditions-in-rx-chains\",\n"
+            + "      \"title\": \"How to include conditions in Rx chains?\"\n"
+            + "    },\n"
+            + "    {\n"
+            + "      \"tags\": [\n"
+            + "        \"rx-java\"\n"
+            + "      ],\n"
+            + "      \"owner\": {\n"
+            + "        \"reputation\": 8,\n"
+            + "        \"user_id\": 3384681,\n"
+            + "        \"user_type\": \"registered\",\n"
+            + "        \"profile_image\": \"https://www.gravatar.com/avatar/c7cf5666f1e47fbae4a9047551f43136?s=128&d=identicon&r=PG&f=1\",\n"
+            + "        \"display_name\": \"offbyone\",\n"
+            + "        \"link\": \"http://stackoverflow.com/users/3384681/offbyone\"\n"
+            + "      },\n"
+            + "      \"is_answered\": true,\n"
+            + "      \"view_count\": 254,\n"
+            + "      \"accepted_answer_id\": 22221410,\n"
+            + "      \"answer_count\": 2,\n"
+            + "      \"score\": 1,\n"
+            + "      \"last_activity_date\": 1467732456,\n"
+            + "      \"creation_date\": 1394040203,\n"
+            + "      \"question_id\": 22204774,\n"
+            + "      \"link\": \"http://stackoverflow.com/questions/22204774/preventing-rapid-clicks-with-rxjava\",\n"
+            + "      \"title\": \"Preventing rapid clicks with RXJava\"\n"
+            + "    }\n"
+            + "  ],\n"
+            + "  \"has_more\": true,\n"
+            + "  \"quota_max\": 300,\n"
+            + "  \"quota_remaining\": 286\n"
+            + "}\n"
+            + "}";
+
+}
