+++ testapplication/src/androidTest/java/com/alterego/stackoverflow/test/data/SearchResponse.java
@@ -1,0 +1,19 @@
+package com.alterego.stackoverflow.test.data;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+import lombok.Getter;
+import lombok.experimental.Accessors;
+
+@Accessors(prefix="m")
+public class SearchResponse {
+
+    @Expose
+    @Getter
+    @SerializedName("items")
+    private List<Question> mQuestions;
+
+}
