+++ testapplication/src/androidTest/java/com/alterego/stackoverflow/test/rules/Repeat.java
@@ -1,0 +1,14 @@
+package com.alterego.stackoverflow.test.rules;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({
+        java.lang.annotation.ElementType.METHOD
+})
+public @interface Repeat {
+
+    int times();
+}
