+++ testapplication/src/main/AndroidManifest.xml
@@ -1,0 +1,13 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="solutions.alterego.testapplication">
+
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+
+    <application android:allowBackup="true"
+                 android:label="Tests"
+                 android:icon="@mipmap/ic_launcher"
+                 android:supportsRtl="true">
+    </application>
+
+</manifest>
